Creep Spread
The creep spreads quite nicely. Seems to be perfectly inline with the
paths that are generated from mapanalyzer. From here, we'll just have to
make sure our queen policy reflects what we want the queens to do. We
can also create different polices to use that maybe we can call at
different times in the game.