Bug Fix: Incorrect Policy
Queens was using it default policy because we weren't passing in parameters correctly. We needed to pass in a boolean for the debug variable. Otherwise, our policy was setting "debug" possibly. I tested that our policy is the one sent in, but need to check how the bot performs on headful.