Creep Spread Done
We reached a point where the creep spread is good enough. We get to about 40% at the 8 minute mark, and are encroaching on the enemies base.

The next key feature for creep spread would be to make sure tumors don't get placed too close to each other, but that might be dependent on the queens-sc2 library.