Creep Refactor
Made a pretty significant refactor. Seems to work fine, except, breaks at iteration 420 with a "Can't await a NoneType" but doesn't show which object is None. So I will have to debug that later.

Added a PathManager also to.. manage our paths. I figured we would be getting and using many different paths, it made
sense to create it.

Also created a setup() function for each manager that gets called from EGbot on_start. That way we can set up all the
managers with their pertinent data.

I think that's it.