Updated queens-sc2; Refactor Complete.
rasper made some edits to his code based on our feedback on the tumor placement. I have added that to our repo. The queens are still placing tumors close to each other but I think that's fine for now.

This refactor is also done. It pieced together surprisingly well. Going to make a pull request to Dev now.